Thue experiment
I invented this at the start of the Quarantine when started reading "Gödel, Escher, Bach: an Eternal Golden Braid", but then found Thue which is very similar to my idea (but mine is totally deterministic). Initially i called it Rulemaker lang, but because of similarity maybe it is Thue's dialect Rulemaker.